+++ forge/src/main/java/cc/cassian/immersiveoverlays/compat/forge/TravelersBackpackCompatImpl.java
@@ -2,7 +2,6 @@ package cc.cassian.immersiveoverlays.compat.forge;
 
 import cc.cassian.immersiveoverlays.overlay.OverlayHelpers;
 import com.tiviacz.travelersbackpack.capability.CapabilityUtils;
-import com.tiviacz.travelersbackpack.inventory.BackpackWrapper;
 import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.items.ItemStackHandler;
 
@@ -10,9 +9,19 @@ public class TravelersBackpackCompatImpl {
 
     public static void checkForImportantAccessories(Player player) {
         if (CapabilityUtils.isWearingBackpack(player)) {
-            BackpackWrapper backpackWrapper = CapabilityUtils.getBackpackWrapper(player);
+            var backpackWrapper =
+            //? if >1.20 {
+             CapabilityUtils.getBackpackWrapper(player);
+            //?} else {
+            /*CapabilityUtils.getBackpackInv(player);
+            *///?}
             if (backpackWrapper != null) {
-                ItemStackHandler inventory = backpackWrapper.inventory;
+                var inventory = backpackWrapper.
+                //? if >1.20 {
+                 inventory; 
+                //?} else {
+                /*getHandler();
+                //?
                 for (int i =0; i < inventory.getSlots(); i++) {
                     OverlayHelpers.isImportantItemOrContainer(inventory.getStackInSlot(i));
                 }
@@ -20,3 +29,4 @@ public class TravelersBackpackCompatImpl {
         }
     }
 }
+*/