+++ src/main/java/cc/cassian/immersiveoverlays/compat/SereneSeasonsCompat.java
@@ -1,10 +1,16 @@
 package cc.cassian.immersiveoverlays.compat;
 
 import net.minecraft.client.multiplayer.ClientLevel;
-import sereneseasons.api.season.SeasonHelper;
+//? if >1.20 {
+  import sereneseasons.api.season.SeasonHelper;
+//?}
 
 public class SereneSeasonsCompat {
     public static String getSeason(ClientLevel level) {
-        return SeasonHelper.getSeasonState(level).getSeason().name();
+        //? if >1.20 {
+            return SeasonHelper.getSeasonState(level).getSeason().name();
+        //?} else {
+        /*return "";
+        *///?}
     }
 }